Many changes, Lights not working, 2d orbits not accounting for inclination of parent body thus 2d orbits not matching mesh position. needs lot more work to clean up.

+++ plugins/environment/mybody/orbit.h
@@ -32,7 +32,6 @@
 #include <string>
 #include <cstdlib>
 #include <cmath>
-#include <ctime>
 #include <vector>
 
 #define SB_AU_KM (149597870.691)  
@@ -94,7 +93,7 @@ public:
 
   csVector3 CirclePoint ( long double angle ) ;  // point on circle at angle 
   csVector3 EllipsePoint ( long double angle ) ; // point on elipse at angle 
-  csVector3 OrbitPointDeg ( long double angle ) ; // point on orbit at angle
+  csVector3 OrbitPointDeg ( float angle ) ; // point on orbit at angle
   csVector3 OrbitPointSec ( long secondspassed ) ;// point on orbit at systime
